+++ HISTORY.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Added
+- [python] seidelWrapper solves 1d LP instead of 2d LP when x_min == x_max and solve_1d > 0
+
 ## 0.5.2 (Nov 19 2022)
 - [cpp] always define all installed symbols.